requirements.txt
numpy==1.23.2
pillow==9.2.0
flask==2.2.2
fastai==2.7.11
tensorflow==2.9.2
resnet-18